safety-tool-parser: move expr_* fns to utils
